Shuffle some types around
