+++ src/RPKI/UniqueId.hs
@@ -9,4 +9,4 @@ thisExecutableVersion :: ExecutableVersion
 thisExecutableVersion = ExecutableVersion $ rpkiProverVersion <> " " <> 
     -- The content is to be updated by the 'src-hash' script 
     -- that calculates hash of the source tree and configuration/build files     
-    "srcHash#eedd2802678da9dbd5ad7ef7c68ac1f54d13863ffc2be379682437f3e6145695#srcHash"
+    "srcHash#f94e5333293dc012d57a137cc80194f340cfb34cf74a932381dc02cb06688a8e#srcHash"
